Refactor resource documentation for clarity and consistency. Update descriptions and schema details across multiple resources, including deployment variable value, policy, relationship rule, and resource. Remove outdated example usages and improve formatting. Adjust internal logic in deployment variable value and policy resources for better readability.

+++ internal/provider/policy_resource.go
@@ -978,7 +978,7 @@ func policyDatadogProviderFromModel(model PolicyDatadogProvider) (api.MetricProv
 		if err != nil {
 			return api.MetricProvider{}, err
 		}
-		seconds := int64(intervalSeconds)
+		seconds := intervalSeconds
 		datadog.IntervalSeconds = &seconds
 	}
 	if selectorValueSet(model.Aggregator) {
